+++ contracts/Profile.sol
@@ -11,10 +11,10 @@ import "./libraries/Errors.sol";
 contract Profile is ERC721URIStorageUpgradeable {
     using Counters for Counters.Counter;
 
-    event URISet(uint256 indexed tokenId, string tokenURI);
+    event URISet(uint indexed tokenId, string tokenURI);
 
     Counters.Counter private _counter;
-    mapping(address => uint256) private _owners;
+    mapping(address => uint) private _owners;
 
     function initialize() public initializer {
         __ERC721_init("Web3 Goals - Profiles", "W3GP");
@@ -31,7 +31,7 @@ contract Profile is ERC721URIStorageUpgradeable {
      * Get uri by owner.
      */
     function getURI(address owner) external view returns (string memory) {
-        uint256 tokenId = _owners[owner];
+        uint tokenId = _owners[owner];
         if (_exists(tokenId)) {
             return tokenURI(tokenId);
         } else {
@@ -48,7 +48,7 @@ contract Profile is ERC721URIStorageUpgradeable {
             // Update counter
             _counter.increment();
             // Mint token
-            uint256 tokenId = _counter.current();
+            uint tokenId = _counter.current();
             _mint(msg.sender, tokenId);
             _owners[msg.sender] = tokenId;
             // Set URI
@@ -63,7 +63,7 @@ contract Profile is ERC721URIStorageUpgradeable {
     /**
      * Set uri.
      */
-    function _setURI(uint256 tokenId, string memory tokenURI) private {
+    function _setURI(uint tokenId, string memory tokenURI) private {
         _setTokenURI(tokenId, tokenURI);
         emit URISet(tokenId, tokenURI);
     }
@@ -74,8 +74,8 @@ contract Profile is ERC721URIStorageUpgradeable {
     function _beforeTokenTransfer(
         address from,
         address to,
-        uint256 firstTokenId,
-        uint256 batchSize
+        uint firstTokenId,
+        uint batchSize
     ) internal virtual override(ERC721Upgradeable) {
         super._beforeTokenTransfer(from, to, firstTokenId, batchSize);
         // Disable transfers except minting
